# api/openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Message Sender API
servers:
  - url: http://localhost:8080
    description: Message sender API for local testing
paths:
  /change-state:
    get:
      summary: Resume or pause the automatic message sender
      description: >
        This endpoint allows you to change the state of the automatic message sender.
        Use this to either resume or pause the sending of messages.
      operationId: changeState
      parameters:
        - name: action
          in: query
          description: Action to perform on the message sender
          required: true
          schema:
            type: string
            enum: [resume, pause]
      responses:
        '200':
          description: State changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
  /sent-messages:
    get:
      summary: Get sent messages
      description: Retrieve a list of messages that have been sent successfully.
      operationId: getSentMessages
      responses:
        '200':
          description: List of sent messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SentMessagesResponse'
components:
  schemas:
    State:
      type: object
      required:
        - running
      properties:
        running:
          type: boolean
          example: true
    Message:
      type: object
      required:
        - id
        - content
        - recipient
        - status
      properties:
        id:
          type: integer
          example: 1
        content:
          type: string
          example: 'Hello!'
        recipient:
          type: string
          example: '+1234567890'
        status:
          type: string
          enum: [sent, unsent, invalid]
          example: sent
        sentAt:
          type: string
          format: date-time
          example: '2025-05-31T10:00:00Z'
    SentMessagesResponse:
      type: array
      items:
            $ref: '#/components/schemas/Message'